Add density-filter and prune-pdbs cli subcommands

+++ src/protein_detective/alphafold/density.py
@@ -1,3 +1,4 @@
+import logging
 from collections.abc import Generator
 from dataclasses import dataclass
 from pathlib import Path
@@ -10,6 +11,9 @@ predicted local distance difference test (pLDDT).
 
 See https://www.ebi.ac.uk/training/online/courses/alphafold/inputs-and-outputs/evaluating-alphafolds-predicted-structures-using-confidence-scores/plddt-understanding-local-confidence/
 """
+
+logger = logging.getLogger(__name__)
+
 
 def find_high_confidence_residues(pdb_file: Path, confidence: float) -> Generator[int]:
     prev_res_index = None
@@ -32,6 +36,9 @@ def filter_out_low_confidence_residues(input_pdb_file: Path, allowed_residues: s
     # elsewhere in the file (SEQRES, TER).
     # TODO do we need to take model/chain into account?
     # now assumes single model and single chain
+    if output_pdb_file.exists():
+        logger.info(f"Output file {output_pdb_file} already exists. Skipping filtering for {input_pdb_file}.")
+        return
     with input_pdb_file.open("r") as input_file, output_pdb_file.open("w") as output_file:
         for line in input_file:
             if line.startswith("ATOM"):
@@ -44,6 +51,13 @@ def filter_out_low_confidence_residues(input_pdb_file: Path, allowed_residues: s
 
 
 @dataclass
+class DensityFilterQuery:
+    confidence: float
+    min_threshold: int
+    max_threshold: int
+
+
+@dataclass
 class DensityFilterResult:
     pdb_file: str
     count: int
@@ -52,17 +66,12 @@ class DensityFilterResult:
 
 
 def filter_on_density(
-    session_dir: Path, confidence: float, min_threshold: int, max_threshold: int
+    alphafold_pdb_files: list[Path], query: DensityFilterQuery, density_filtered_dir: Path
 ) -> Generator[DensityFilterResult]:
-    download_dir = session_dir / "downloads"
-    density_filtered_dir = session_dir / "density_filtered"
-    density_filtered_dir.mkdir(parents=True, exist_ok=True)
-
-    # TODO get list of files from filesystem or from session db?
-    for pdb_file in download_dir.glob("AF-*.pdb"):
-        residues = set(find_high_confidence_residues(pdb_file, confidence))
+    for pdb_file in alphafold_pdb_files:
+        residues = set(find_high_confidence_residues(pdb_file, query.confidence))
         count = len(residues)
-        if count < min_threshold or count > max_threshold:
+        if count < query.min_threshold or count > query.max_threshold:
             yield DensityFilterResult(
                 pdb_file=pdb_file.name,
                 count=count,
